+++ Wrappers/Python/cil/framework/vector_geometry.py
@@ -112,7 +112,7 @@ class VectorGeometry:
                     half_dtype = numpy.real(complex_example).dtype
                     r = global_rng.random(size=self.shape, dtype=half_dtype) + 1j * global_rng.random(size=self.shape, dtype=half_dtype)
                 else:
-                    r = global_rng.random(size=self.shape, dtype=out.dtype)
+                    r = global_rng.random(size=self.shape, dtype=dtype)
 
             elif value == FillType.RANDOM_INT:
                 max_value = kwargs.get('max_value', 100)
